refactor(auth): update onboarding types and file naming
- Update professional field type to UserValidatorsCreateUserSchemaProfessionalFieldItem
- Fix business field label translation
- Remove unnecessary array indexing in form submission
- Rename step components to kebab-case for consistency
