+++ widgets/auth/ui/steps/index.ts
@@ -1,3 +1,3 @@
-export { PersonalInfoStep } from './PersonalInfoStep';
-export { InterestsStep } from './InterestsStep';
-export { ReferralSourceStep } from './ReferralSourceStep';
+export { PersonalInfoStep } from './personal-info-step';
+export { InterestsStep } from './interests-step';
+export { ReferralSourceStep } from './referral-source-step';
