+++ widgets/auth/ui/steps/interests-step.tsx
