+++ widgets/auth/ui/steps/personal-info-step.tsx
