+++ widgets/auth/ui/steps/referral-source-step.tsx
